requirements.txt
Django==1.8.3
Django==1.8.3
django-autoslug==1.8.0
